+++ Runtime/BitUtils.cs
@@ -1,4 +1,5 @@
 using System.Runtime.CompilerServices;
+using Unity.Mathematics;
 
 namespace Hydrogen.Maths
 {
@@ -8,6 +9,24 @@ namespace Hydrogen.Maths
     /// </summary>
     public static class BitUtils
     {
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool4x2 ToBool4x2(this byte value)
+        {
+            var lo = new bool4((value & 0x01) != 0, (value & 0x02) != 0, (value & 0x04) != 0, (value & 0x08) != 0);
+            var hi = new bool4((value & 0x10) != 0, (value & 0x20) != 0, (value & 0x40) != 0, (value & 0x80) != 0);
+
+            return new bool4x2(lo, hi);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool4x4 ToBool4x4(this ushort value)
+        {
+            var lo = ToBool4x2((byte)(value & 0x00FF));
+            var hi = ToBool4x2((byte)((value & 0xFF00) >> 8));
+
+            return new bool4x4(lo.c0, lo.c1, hi.c0, hi.c1);
+        }
+
         /// <summary>
         /// Derived from: http://graphics.stanford.edu/~seander/bithacks.html#SwappingBitsXO
         /// but using a constant 1 bit swap length
@@ -19,8 +38,8 @@ namespace Hydrogen.Maths
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static byte Swap(byte b, byte i, byte j)
         {
-            byte x = (byte) (((b >> i) ^ (b >> j)) & 1);
-            byte r = (byte) (b ^ ((x << i) | (x << j)));
+            var x = (byte) (((b >> i) ^ (b >> j)) & 1);
+            var r = (byte) (b ^ ((x << i) | (x << j)));
             return r;
         }
         
@@ -37,8 +56,8 @@ namespace Hydrogen.Maths
         {
             unchecked
             {
-                byte x = (byte) (((b >> i) ^ (b >> j)) & ((1 << n) - 1)); // XOR temporary
-                byte r = (byte) (b ^ ((x << i) | (x << j)));
+                var x = (byte) (((b >> i) ^ (b >> j)) & ((1 << n) - 1)); // XOR temporary
+                var r = (byte) (b ^ ((x << i) | (x << j)));
                 return r;
             }
         }
@@ -54,8 +73,8 @@ namespace Hydrogen.Maths
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static ushort Swap(ushort b, byte i, byte j)
         {
-            ushort x = (ushort) (((b >> i) ^ (b >> j)) & 1);
-            ushort r = (ushort) (b ^ ((x << i) | (x << j)));
+            var x = (ushort) (((b >> i) ^ (b >> j)) & 1);
+            var r = (ushort) (b ^ ((x << i) | (x << j)));
             return r;
         }
 
@@ -72,8 +91,8 @@ namespace Hydrogen.Maths
         {
             unchecked
             {
-                ushort x = (ushort) (((b >> i) ^ (b >> j)) & ((1 << n) - 1)); // XOR temporary
-                ushort r = (ushort) (b ^ ((x << i) | (x << j)));
+                var x = (ushort) (((b >> i) ^ (b >> j)) & ((1 << n) - 1)); // XOR temporary
+                var r = (ushort) (b ^ ((x << i) | (x << j)));
                 return r;
             }
         }
@@ -89,8 +108,8 @@ namespace Hydrogen.Maths
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static uint Swap(uint b, byte i, byte j)
         {
-            uint x = ((b >> i) ^ (b >> j)) & 1u;
-            uint r = b ^ ((x << i) | (x << j));
+            var x = ((b >> i) ^ (b >> j)) & 1u;
+            var r = b ^ ((x << i) | (x << j));
             return r;
         }
 
@@ -107,8 +126,8 @@ namespace Hydrogen.Maths
         {
             unchecked
             {
-                uint x = ((b >> i) ^ (b >> j)) & ((1u << n) - 1u); // XOR temporary
-                uint r = b ^ ((x << i) | (x << j));
+                var x = ((b >> i) ^ (b >> j)) & ((1u << n) - 1u); // XOR temporary
+                var r = b ^ ((x << i) | (x << j));
                 return r;
             }
         }
@@ -124,8 +143,8 @@ namespace Hydrogen.Maths
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static ulong Swap(ulong b, byte i, byte j)
         {
-            ulong x = ((b >> i) ^ (b >> j)) & 1ul;
-            ulong r = b ^ ((x << i) | (x << j));
+            var x = ((b >> i) ^ (b >> j)) & 1ul;
+            var r = b ^ ((x << i) | (x << j));
             return r;
         }
 
@@ -142,8 +161,8 @@ namespace Hydrogen.Maths
         {
             unchecked
             {
-                ulong x = ((b >> i) ^ (b >> j)) & ((1ul << n) - 1ul); // XOR temporary
-                ulong r = b ^ ((x << i) | (x << j));
+                var x = ((b >> i) ^ (b >> j)) & ((1ul << n) - 1ul); // XOR temporary
+                var r = b ^ ((x << i) | (x << j));
                 return r;
             }
         }