+++ Tests/Runtime/TestConstants.cs
@@ -1,0 +1,17 @@
+namespace Hydrogen.Maths.Tests
+{
+    public static class TestConstants
+    {
+        public const byte ByteBefore = 0x0F;
+        public const byte ByteAfter = 0xF0;
+
+        public const ushort UShortBefore = 0x00FF;
+        public const ushort UShortAfter = 0xFF00;
+
+        public const uint UIntBefore = 0x0000FFFF;
+        public const uint UIntAfter = 0xFFFF0000;
+
+        public const ulong ULongBefore = 0x00000000FFFFFFFF;
+        public const ulong ULongAfter = 0xFFFFFFFF00000000;
+    }
+}